% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_points_in_ellipsoid.R
\name{extract_points_in_ellipsoid}
\alias{extract_points_in_ellipsoid}
\title{Extract points that fall inside a given ellipsoid}
\usage{
extract_points_in_ellipsoid(ell, points)
}
\arguments{
\item{ell}{An "ellipsoid" object from get_parametric_ellipsoid().}

\item{points}{A numeric matrix of points to test (rows are points, columns are dimensions).}
}
\value{
A matrix containing only the rows (points) that fall inside the ellipsoid.
}
\description{
Extract points that fall inside a given ellipsoid
}
