% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrences_3d.R
\name{plot_occurrences_3d}
\alias{plot_occurrences_3d}
\title{Plot Sampled Occurrences in 3D Environmental Space}
\usage{
plot_occurrences_3d(occurrences, env_rasters, niche, n_bg = 20000, ...)
}
\arguments{
\item{occurrences}{A data frame of occurrence points from `sample_from_env()`.}

\item{env_rasters}{The `SpatRaster` object used to generate the occurrences.}

\item{niche}{The `ellipsoid` object used to generate the occurrences.}

\item{n_bg}{The number of background points to sample for visualization.}

\item{...}{Additional arguments passed to `plot_ly`.}
}
\value{
An interactive `plotly` object.
}
\description{
Creates an interactive 3D plot showing the sampled occurrence points
relative to the available background environments and the niche boundary.
}
\seealso{
Other ellipsoid functions: 
\code{\link{build_ellipsoid}()},
\code{\link{generate_suitability}()},
\code{\link{plot_niche_3d}()},
\code{\link{points_in_ellipsoid}()},
\code{\link{sample_points_niche}()}
}
\concept{ellipsoid functions}
