% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_in_ellipsoid.R
\name{points_in_ellipsoid}
\alias{points_in_ellipsoid}
\title{Extract Points Within an Ellipsoid}
\usage{
points_in_ellipsoid(ell, points)
}
\arguments{
\item{ell}{An `ellipsoid` object created by `build_ellipsoid()`.}

\item{points}{A matrix or data frame of points to test. The column order
must match the order of variables in the ellipsoid's center vector.}
}
\value{
A matrix containing the rows from `points` that are inside the ellipsoid.
}
\description{
Filters a set of points, returning only those that fall within the boundary
of a given ellipsoid, based on the squared Mahalanobis distance.
}
\seealso{
Other ellipsoid functions: 
\code{\link{build_ellipsoid}()},
\code{\link{generate_suitability}()},
\code{\link{plot_niche_3d}()},
\code{\link{plot_occurrences_3d}()},
\code{\link{sample_points_niche}()}
}
\concept{ellipsoid functions}
