% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ellipsoid.R
\name{build_ellipsoid}
\alias{build_ellipsoid}
\title{Build a 2D or 3D Ellipsoid Object}
\usage{
build_ellipsoid(
  center = c(x = 0, y = 0),
  axes = c(1, 1),
  angles = 0,
  n_points = 100
)
}
\arguments{
\item{center}{A named numeric vector of length 2 or 3 defining the ellipsoid's
center (e.g., `c(Temp = 15, Precip = 800)`).}

\item{axes}{A numeric vector with the same length as `center`, defining the
semi-axis lengths *before* rotation.}

\item{angles}{Numeric vector of rotation angles in radians. For 2D, a single
angle; for 3D, a vector of three angles `c(ax, ay, az)`.}

\item{n_points}{Integer; the resolution for the plotted surface.}
}
\value{
An object of class `ellipsoid`.
}
\description{
Creates a structured object representing a rotated ellipsoid. This object
contains all necessary matrices for geometric and statistical calculations,
as well as surface points for plotting.
}
\seealso{
Other ellipsoid functions: 
\code{\link{generate_suitability}()},
\code{\link{plot_niche_3d}()},
\code{\link{plot_occurrences_3d}()},
\code{\link{points_in_ellipsoid}()},
\code{\link{sample_points_niche}()}
}
\concept{ellipsoid functions}
