% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_niche_3d.R
\name{plot_niche_3d}
\alias{plot_niche_3d}
\title{Plot a 3D Niche in Environmental Space}
\usage{
plot_niche_3d(
  env_data,
  niche,
  n_bg = 50000,
  n_suitable = 10000,
  title = "Ellipsoid Niche in 3D Environmental Space"
)
}
\arguments{
\item{env_data}{A data frame containing columns for the environmental variables
(e.g., "AnnualTemp", "AnnualPrecip", "TempSeasonality") and a `suitability`
column (1 for suitable, 0 for unsuitable).}

\item{niche}{An `ellipsoid` object created by `build_ellipsoid()`. It must
be a 3D niche.}

\item{n_bg}{An integer specifying the number of background (unsuitable) points
to sample for the plot.}

\item{n_suitable}{An integer specifying the number of suitable points to sample.}

\item{title}{A character string for the plot's main title.}
}
\value{
An interactive `plotly` object.
}
\description{
Creates an interactive 3D plot showing the species' ellipsoidal niche
relative to the available background environments.
}
\examples{
# This is a conceptual example. You would run this after creating
# 'bio_df_with_suitability' and 'species_niche' as shown in the README.

# plot_3d <- plot_niche_3d(
#   env_data = bio_df_with_suitability,
#   niche = species_niche,
#   n_bg = 50000,
#   n_suitable = 10000
# )
#
# plot_3d
}
\seealso{
Other ellipsoid functions: 
\code{\link{build_ellipsoid}()},
\code{\link{generate_suitability}()},
\code{\link{points_in_ellipsoid}()}
}
\concept{ellipsoid functions}
